requirements.txt
Flask==2.0.1
Pillow==9.0.0
numpy==1.24.3
opencv-python==4.8.0.76 